+++ realtwin/__init__.py
@@ -1,6 +1,6 @@
 
 ##############################################################################
-# Copyright (c) 2024, Oak Ridge National Laboratory                          #
+# Copyright (c) 2024-, Oak Ridge National Laboratory                          #
 # All rights reserved.                                                       #
 #                                                                            #
 # This file is part of RealTwin and is distributed under a GPL               #
@@ -18,8 +18,13 @@ from realtwin.util_lib.create_venv import venv_create, venv_delete
 from realtwin.func_lib._a_install_simulator.check_sim_env import (is_sumo_installed,
                                                                   is_vissim_installed,
                                                                   is_aimsun_installed)
-from realtwin.util_lib.create_config import prepare_config_file
+from realtwin.util_lib.create_config import prepare_realtwin_configs
 from realtwin.util_lib.download_elevation_tif import download_elevation_tif_by_bbox
+
+# Autonomous vehicle simulation
+from realtwin.autonomous_veh.sim_av import SimAV, prepare_av_configs
+# from realtwin.autonomous_veh._generate_loop_detector import generate_sumo_loop_detector_add_xml
+# from realtwin.autonomous_veh._load_config import load_av_configs
 
 __version__ = '0.1.0'
 
@@ -33,10 +38,15 @@ __all__ = [
 
     # util_lib.create_venv
     'venv_create', 'venv_delete',
-    'prepare_config_file',
+    'prepare_realtwin_configs',
     'download_elevation_tif_by_bbox',
 
     # func_lib
+
+    # autonomous_veh
+    'SimAV', 'prepare_av_configs',
+    # 'generate_sumo_loop_detector_add_xml',
+    # 'load_av_configs',
 ]